requirements.txt
playwright>=1.52.0
google-generativeai>=0.8.5
python-dotenv>=1.0.1
cryptography>=43.0.1
pydantic>=2.9.2
